+++ starsky/starsky.foundation.sync/WatcherInterfaces/IQueueProcessor.cs
@@ -1,0 +1,10 @@
+using System.IO;
+
+namespace starsky.foundation.sync.WatcherInterfaces
+{
+	public interface IQueueProcessor
+	{
+		void QueueInput(string filepath, string toPath,
+			WatcherChangeTypes changeTypes);
+	}
+}